% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayCAST-package.R
\docType{package}
\name{baycast-package}
\alias{BayCAST}
\alias{BayCAST-package}
\alias{baycast-package}
\title{baycast: Bayesian Dirichlet-Process AFT with Factor-Analytic Spike-and-Slab Loadings}
\description{
A Bayesian semiparametric model for survival data with high-dimensional
covariates. The latent factor model is shared across a Dirichlet-process
mixture of accelerated-failure-time components; spike-and-slab variable
selection (SSVS) is applied to the loading matrix.
}
\section{Main function}{

\code{\link[=baycast]{baycast()}} is the user-facing entry point. Use \code{\link[=baycast_control]{baycast_control()}} to
adjust advanced hyperparameters.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jihyun1114/BayCAST}
  \item Report bugs at \url{https://github.com/jihyun1114/BayCAST/issues}
}

}
\author{
\strong{Maintainer}: Jihyun Kim \email{kim02688@umn.edu}

}

}
