% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior_z}
\alias{posterior_z}
\title{Posterior mode of cluster assignment}
\usage{
posterior_z(x)
}
\arguments{
\item{x}{A fitted \code{bdpaft} object.}
}
\value{
Integer vector of length \code{n} (number of subjects).
}
\description{
For each subject, returns the most frequent cluster label in the
posterior draws.
}
